st/django_project/django/pyproject.toml
[tool.poetry]
name = "proj"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
# readme = "README.md"
packages = [{include = "proj"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
daphne = "^4.0.0"
djangorestframework = "^3.14.0"
pytz = "^2022.7.1"
psycopg2-binary = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
